python-app/requirements.txt
flask>=2.3.3
pyyaml>=6.0.1
argon2-cffi>=23.1.0
cryptography>=41.0.4
typing-extensions>=4.8.0
